feat: tsx convert
