+++ src/components/solution.tsx
@@ -1,0 +1,142 @@
+import {
+  ArrowPathIcon,
+  BookmarkSquareIcon,
+  CalendarIcon,
+  ChartBarIcon,
+  CursorArrowRaysIcon,
+  LifebuoyIcon,
+  PhoneIcon,
+  PlayIcon,
+  ShieldCheckIcon,
+  Squares2X2Icon,
+} from '@heroicons/react/24/outline';
+
+const solutions = [
+  {
+    name: 'Docs',
+    description:
+      'Get a better understanding of where your traffic is coming from.',
+    href: 'docs',
+    icon: ChartBarIcon,
+  },
+  {
+    name: '404',
+    description: 'Speak directly to your customers in a more meaningful way.',
+    href: '404',
+    icon: CursorArrowRaysIcon,
+  },
+  {
+    name: 'Info',
+    description: "Your customers' data will be safe and secure.",
+    href: 'info',
+    icon: ShieldCheckIcon,
+  },
+  {
+    name: 'Integrations',
+    description: "Connect with third-party tools that you're already using.",
+    href: '#',
+    icon: Squares2X2Icon,
+  },
+  {
+    name: 'Automations',
+    description:
+      'Build strategic funnels that will drive your customers to convert',
+    href: '#',
+    icon: ArrowPathIcon,
+  },
+];
+const callsToAction = [
+  { name: 'Watch Demo', href: '#', icon: PlayIcon },
+  { name: 'Contact Sales', href: '#', icon: PhoneIcon },
+];
+const resources = [
+  {
+    name: 'Help Center',
+    description:
+      'Get all of your questions answered in our forums or contact support.',
+    href: '#',
+    icon: LifebuoyIcon,
+  },
+  {
+    name: 'Guides',
+    description:
+      'Learn how to maximize our platform to get the most out of it.',
+    href: '#',
+    icon: BookmarkSquareIcon,
+  },
+  {
+    name: 'Events',
+    description:
+      'See what meet-ups and other events we might be planning near you.',
+    href: '#',
+    icon: CalendarIcon,
+  },
+  {
+    name: 'Security',
+    description: 'Understand how we take your privacy seriously.',
+    href: '#',
+    icon: ShieldCheckIcon,
+  },
+];
+const recentPosts = [
+  { id: 1, name: 'Boost your conversion rate', href: '#' },
+  {
+    id: 2,
+    name: 'How to use search engine optimization to drive traffic to your site',
+    href: '#',
+  },
+  { id: 3, name: 'Improve your customer experience', href: '#' },
+];
+
+const categories = {
+  Recent: [
+    {
+      id: 1,
+      title: 'Does drinking coffee make you smarter?',
+      date: '5h ago',
+      commentCount: 5,
+      shareCount: 2,
+    },
+    {
+      id: 2,
+      title: "So you've bought coffee... now what?",
+      date: '2h ago',
+      commentCount: 3,
+      shareCount: 2,
+    },
+  ],
+  Popular: [
+    {
+      id: 1,
+      title: 'Is tech making coffee better or worse?',
+      date: 'Jan 7',
+      commentCount: 29,
+      shareCount: 16,
+    },
+    {
+      id: 2,
+      title: 'The most innovative things happening in coffee',
+      date: 'Mar 19',
+      commentCount: 24,
+      shareCount: 12,
+    },
+  ],
+  Trending: [
+    {
+      id: 1,
+      title: 'Ask Me Anything: 10 answers to your questions about coffee',
+      date: '2d ago',
+      commentCount: 9,
+      shareCount: 5,
+    },
+    {
+      id: 2,
+      title: "The worst advice we've ever heard about coffee",
+      date: '4d ago',
+      commentCount: 1,
+      shareCount: 2,
+    },
+  ],
+};
+
+export { solutions, recentPosts, resources, callsToAction, categories };
